pyproject.toml
[project]
name = "mcpflux"
version = "1.0.0"
description = "Spreadsheet MCP Agent - Query CSV/Excel files with natural language using Claude"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=0.1.0",
    "fastmcp>=0.0.10",
    "pandas>=2.0.0",
    "duckdb>=0.9.0",
    "openpyxl>=3.0.0",
    "python-dotenv>=1.0.0",
    "anthropic>=0.7.0",
    "langchain>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langsmith>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "ruff>=0.0.250",
]

[tool.uv]
package = true

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.ruff]
line-length = 100
target-version = "py310"